Rename all extension `::new()` constructors to `::load()` (#1021)
These functions don't create any refcounted objects or similar, but
exclusively load the function pointers for extensions which is not a
very costly but at least reentrant operation.

Signify this by repeating the underlying `*Fn::load()` name; the only
additional thing these do is receive an entry/instance/device to load
the pointers from and sometimes store the object handle to reuse in
function calls later for convenience.